Android emulator done

+++ build.sh
@@ -23,7 +23,11 @@ elif [[ "${1:-}" == "web" ]]; then
 	echo "Done! - Check in browser"
 
 elif [[ "${1:-}" == "android" ]]; then
-	docker run -e TMDB_KEY -v ~/.pub-cache:/pub-cache -e PUB_CACHE=/pub-cache -v `pwd`:/src -w /src --rm fmv_tools bash _ops/build.android.sh
+	if type docker; then
+		docker run -e TMDB_KEY -v ~/.pub-cache:/pub-cache -e PUB_CACHE=/pub-cache -v `pwd`:/src -w /src --rm fmv_tools bash _ops/build.android.sh
+	else
+		PUB_CACHE=~/.pub-cache bash _ops/build.android.mac.sh
+	fi
 
 elif [[ "${1:-}" == "ios" ]]; then
 	bash _ops/build.ios.sh
